requirements.txt
flask>=3.1
psycopg[binary]>=3.1
python-dotenv>=1.0
msal>=1.20
